Implemented createFolder lambda function
