+++ .gitignore
@@ -4,3 +4,4 @@ credentials.json
 node_modules/
 hasura/hasura.exe
 *.graphqlconfig
+token_drive_v3.pickle
