--- backend/googleapi/requirements.txt
@@ -1,33 +1,0 @@
-cachetools==4.2.0
-certifi==2020.12.5
-chardet==3.0.4
-click==7.1.2
-Flask==1.1.2
-Flask-Cors==3.0.9
-Flask-MySQLdb==0.2.0
-Flask-SQLAlchemy==2.4.4
-google-api-core==1.24.0
-google-api-python-client==1.12.8
-google-auth==1.24.0
-google-auth-httplib2==0.0.4
-google-auth-oauthlib==0.4.2
-googleapis-common-protos==1.52.0
-httplib2==0.18.1
-idna==2.10
-itsdangerous==1.1.0
-Jinja2==2.11.2
-MarkupSafe==1.1.1
-mysqlclient==2.0.2
-oauthlib==3.1.0
-protobuf==3.14.0
-pyasn1==0.4.8
-pyasn1-modules==0.2.8
-pytz==2020.4
-requests==2.25.0
-requests-oauthlib==1.3.0
-rsa==4.6
-six==1.15.0
-SQLAlchemy==1.3.20
-uritemplate==3.0.1
-urllib3==1.26.2
-Werkzeug==1.0.1
